Make sure view UI colors continue to be updated
After color floating sheet is refactored into Compose, its binder will
no longer be called. Move code block for binding view UI colors to the
parent binder so it continues to be binded. This change does not affect
behavior when the flag is off and fixes view color binding when the flag
is on.

Flag: com.android.wallpaper.color_picker_update_flag
Bug: 450071500
Test: manually verified with flag on and off
Change-Id: I787ed54ccf04a53961f03cc8492ae348eda979ea