Adjust color package for theme service flag
When setting color to secure settings, make sure color option package
always includes system palette if the theme service flag is on. Ideally,
the theme service should be used to set color options, but this code
path could be used if the color picker update flag is off and theme
service flag is on.

Flag: android.server.enable_theme_service
Bug: 461547295
Test: unit tests
Test: manually verified logs when applying color
Change-Id: Ic71def09acd1b50b0fbcf79b8ff77d73a783ffb4